Bring back test_two_nodes with current consensus
