+++ core/txflow/src/dag/mod.rs
@@ -2,7 +2,7 @@ mod message;
 mod reporter;
 
 use primitives::signature::DEFAULT_SIGNATURE;
-use primitives::traits::{Payload, WitnessSelector};
+use primitives::traits::{Payload, WitnessSelector, ConsensusBlockBody};
 use primitives::types::*;
 
 use std::cell::RefCell;
@@ -21,7 +21,7 @@ use typed_arena::Arena;
 /// the references never outlive the instances.
 pub struct DAG<
     'a,
-    P: 'a + Payload,
+    P: 'a + Payload + Default,
     W: 'a + WitnessSelector,
     M: 'a + MisbehaviorReporter = NoopMisbehaviorReporter,
 > {
@@ -47,7 +47,7 @@ pub struct DAG<
     misbehavior: Box<RefCell<M>>,
 }
 
-impl<'a, P: 'a + Payload, W: WitnessSelector, M: 'a + MisbehaviorReporter> DAG<'a, P, W, M> {
+impl<'a, P: 'a + Payload + Default, W: WitnessSelector, M: 'a + MisbehaviorReporter> DAG<'a, P, W, M> {
     pub fn new(owner_uid: UID, beacon_block_index: u64, starting_epoch: u64, witness_selector: &'a W) -> Self {
         DAG {
             owner_uid,
@@ -142,7 +142,7 @@ impl<'a, P: 'a + Payload, W: WitnessSelector, M: 'a + MisbehaviorReporter> DAG<'
         publishable.sort_by(|(epoch1, _), (epoch2, _)| epoch1.cmp(epoch2));
 
         // Returned consensuses. Lowest epoch first.
-        let res = vec![];
+        let mut res = vec![];
         // TODO(#125) Currently this goes through without beacon chain consensus. Once we have a
         // beacon chain consensus the epoch will be used.
         for (epoch, group) in publishable {
@@ -151,10 +151,14 @@ impl<'a, P: 'a + Payload, W: WitnessSelector, M: 'a + MisbehaviorReporter> DAG<'
                 .iter().next().expect("At least one message expected.");
             let mut parents = vec![];
             self.collect_parents(repr, &mut parents);
-//            res.push(ConsensusBlockBody {
-//                messages: parents.iter().map(|m| m.data.clone()).collect(),
-//                beacon_block_index: self.beacon_block_index
-//            });
+            let mut payload = P::default();
+            for m in parents.iter() {
+                payload.union_update(m.data.body.payload.clone());
+            }
+            res.push(ConsensusBlockBody {
+                payload,
+                beacon_block_index: self.beacon_block_index
+            });
             self.published_epochs.insert(*epoch);
         }
         res
