+++ node/network/Cargo.toml
@@ -19,8 +19,6 @@ serde_json = "1.0"
 tokio = "0.1"
 tokio-codec = "0.1"
 tokio-serde-cbor = "0.3"
-actix = "0.7.9"
-actix-net = "0.2.6"
 
 beacon = { path = "../../node/beacon" }
 chain = { path = "../../core/chain" }
